+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.12] - 2024-08-28
+
+### Added 
+
+- Exceptions file, with invaled move, invaled move rook and invaled move no piece
+- To turn method in chess, added the exception for invaled move
+- Set piece method in board class
+- Possible vertical up and down positions for rook
+
+### Known Issues
+
+- Test for the rook str doesnt pass
+- Valid positions for rook unfinished
+- Test for vertical down move for rook doesnt pass
+
 ## [0.0.11] - 2024-08-27
 
 ### Changed